Add alternate comicbook and Mobipocket-base mimetypes
The list of officially recognized IANA Media Types is not comprehensive,
nor universally followed. In particular, I see considerable variation
among Web servers on which media type they claim a linked downloadable
file to be. So I figure it's better for Lector to claim the alternative
media types too, so that users will see Lector as an option when
clicking to open local and online comics and ebooks.